chore: restructure project folders and add path aliases
