+++ src/App.tsx
@@ -1,4 +1,4 @@
-import { AuthProvider } from '@context'
+import { AuthProvider } from '@provider'
 import { Layout } from '@components'
 import { AppRoutes } from '@routes'
