+++ src/context/AuthContext.tsx
@@ -1,35 +1,4 @@
-import { useState, useEffect } from 'react'
-import type { ReactNode } from 'react'
-import { AuthContext } from '@types'
-import type { IUser } from '@types'
+import { createContext } from 'react'
+import type { IAuthContext } from '@types'
 
-export const AuthProvider = ({ children }: { children: ReactNode }) => {
-  const [user, setUser] = useState<IUser | null>(null)
-
-  // Load user from localStorage on mount
-  useEffect(() => {
-    const storedUser = localStorage.getItem('user')
-    if (storedUser) {
-      try {
-        setUser(JSON.parse(storedUser))
-      } catch (error) {
-        console.error('Error parsing stored user:', error)
-        localStorage.removeItem('user')
-      }
-    }
-  }, [])
-
-  const login = (user: IUser) => {
-    setUser(user)
-    localStorage.setItem('user', JSON.stringify(user))
-  }
-
-  const logout = () => {
-    setUser(null)
-    localStorage.removeItem('user')
-  }
-
-  return (
-    <AuthContext.Provider value={{ user, setUser, login, logout }}>{children}</AuthContext.Provider>
-  )
-}
+export const AuthContext = createContext<IAuthContext | undefined>(undefined)
