+++ src/provider/AuthProvider.tsx
@@ -1,0 +1,35 @@
+import { useState, useEffect } from 'react'
+import type { ReactNode } from 'react'
+import type { IUser } from '@types'
+import { AuthContext } from '@context'
+
+export const AuthProvider = ({ children }: { children: ReactNode }) => {
+  const [user, setUser] = useState<IUser | null>(null)
+
+  // Load user from localStorage on mount
+  useEffect(() => {
+    const storedUser = localStorage.getItem('user')
+    if (storedUser) {
+      try {
+        setUser(JSON.parse(storedUser))
+      } catch (error) {
+        console.error('Error parsing stored user:', error)
+        localStorage.removeItem('user')
+      }
+    }
+  }, [])
+
+  const login = (user: IUser) => {
+    setUser(user)
+    localStorage.setItem('user', JSON.stringify(user))
+  }
+
+  const logout = () => {
+    setUser(null)
+    localStorage.removeItem('user')
+  }
+
+  return (
+    <AuthContext.Provider value={{ user, setUser, login, logout }}>{children}</AuthContext.Provider>
+  )
+}
