+++ src/types/auth.types.ts
@@ -1,5 +1,3 @@
-import { createContext } from 'react'
-
 export interface IUser {
   id: string
   name: string
@@ -13,5 +11,3 @@ export interface IAuthContext {
   login: (user: IUser) => void
   logout: () => void
 }
-
-export const AuthContext = createContext<IAuthContext | undefined>(undefined)
