+++ tsconfig.app.json
@@ -32,7 +32,8 @@
       "@locales": ["./src/locales"],
       "@mocks": ["./src/mocks"],
       "@hooks": ["./src/hooks"],
-      "@context": ["./src/context"]
+      "@context": ["./src/context"],
+      "@provider": ["./src/provider"]
     }
   },
   "include": ["src"]
